Update to README, file structure.
